refactor(array): extract shared element row logic into useArrayElementRows hook
AddKeyArray and ArrayAddElementForm shared near-identical copies of addField,
clearElement, onClickRemove, isClearDisabled, handleChange, the focus-on-add
effect, and the integer index validation. Extract all of it into a single
useArrayElementRows hook (browser/hooks/useArrayElementRows.ts).

The hook accepts emptyIndexValue to handle the one intentional difference
between the two forms: AddKeyArray resets to '0' (sensible default for create),
while ArrayAddElementForm resets to '' (user must supply an explicit index).
